+++ src/main/kotlin/org/appjam/smashing/domain/user/dto/response/CountsResult.kt
@@ -1,0 +1,6 @@
+package org.appjam.smashing.domain.user.dto.response
+
+data class CountsResult(
+    val ratingCounts: UserRecentReviewSummaryResponse.RatingCounts,
+    val tagCounts: UserRecentReviewSummaryResponse.TagCounts,
+)
