+++ src/main/kotlin/org/appjam/smashing/domain/user/dto/response/UserWithActiveProfile.kt
@@ -1,0 +1,9 @@
+package org.appjam.smashing.domain.user.dto.response
+
+import org.appjam.smashing.domain.user.entity.User
+import org.appjam.smashing.domain.user.entity.UserSportProfile
+
+data class UserWithActiveProfile(
+    val user: User,
+    val activeProfile: UserSportProfile,
+)
